requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
packaging==26.1
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.12.1
python-decouple==3.8
pytz==2026.1.post1
reportlab==4.0.7
setuptools==82.0.1
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.6.0
